requirements.txt
opencv-python
sklearn
torchvision==0.5.0
tensorflow==1.15
matplotlib
seaborn
